separate Series timezone tests
